+++ Wireshark/pip_download.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+pip install --download ./pip pyshark
