+++ Wireshark/pip_offline_install.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+pip install --no-index --find-links ./pip pyshark
